pyproject.toml
[tool.poetry]
name = "sporz-quicksheet"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "sporz_quicksheet"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.4.1"
rich = "^13.5.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.15.0"
tabulate = "^0.9.0"
pexpect = "^4.8.0"
devtools = "^0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
